Merge/2.0

+++ molgenis-data-vcf/src/main/java/org/molgenis/data/vcf/VcfAttributes.java
@@ -1,13 +1,13 @@
 package org.molgenis.data.vcf;
 
 import static java.util.Objects.requireNonNull;
-import static org.molgenis.MolgenisFieldTypes.LONG;
+import static org.molgenis.MolgenisFieldTypes.INT;
 import static org.molgenis.MolgenisFieldTypes.STRING;
 import static org.molgenis.MolgenisFieldTypes.TEXT;
 import static org.molgenis.data.vcf.VcfRepository.DEFAULT_ATTRIBUTE_DESCRIPTION;
 
-import org.molgenis.data.meta.AttributeMetaData;
-import org.molgenis.data.meta.AttributeMetaDataFactory;
+import org.molgenis.data.meta.model.AttributeMetaData;
+import org.molgenis.data.meta.model.AttributeMetaDataFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -48,7 +48,7 @@ public class VcfAttributes
 
 	public AttributeMetaData getPosAttribute()
 	{
-		return attributeMetaDataFactory.create().setName(POS).setDataType(LONG).setAggregatable(true).setNillable(false)
+		return attributeMetaDataFactory.create().setName(POS).setDataType(INT).setAggregatable(true).setNillable(false)
 				.setDescription("The position on the chromosome which the variant is observed");
 	}
